requirements.txt
streamlit>=1.28.0
youtube-transcript-api>=0.6.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-pinecone>=0.1.0
langchain-groq>=0.1.0
pinecone-client>=3.0.0
sentence-transformers>=2.2.2
requests>=2.25.0
python-dotenv>=1.0.0
protobuf<3.21.0
